refactor: moved helpers to actions package and sync with latest progresses
